Refactor VJSON module into package. (#7)
The tutor_recon.util.vjson module had become a monster. It was time to break it out into its own package.

The interface is preserved by re-exporting names in tutor_recon.util.vjson.__init__.
