--- tutor_recon/util/vjson.py
@@ -1,452 +1,0 @@
-"""A custom JSON decoder and associated utilities."""
-
-from abc import ABC, abstractclassmethod, abstractmethod
-import json
-from json import JSONDecoder, JSONEncoder
-from collections import MutableMapping
-from typing import Literal, Optional, Type, Union
-from pathlib import Path
-from copy import copy
-
-from tutor_recon.util.misc import WrappedDict, brief, set_nested, walk_dict
-
-MARKER = "$"
-
-JSON_T = Union[str, int, float, bool, list, dict]
-"""A type which can be represented in JSON."""
-
-NOTHING = object()
-"""
-Used in place of `None` in this module to indicate a parameter without a value,
-since `None` maps to `null`, a valid JSON primitive.
-"""
-
-IGNORE = object()
-"""
-When produced as a value, indicates that the entire keypair should be ignored.
-"""
-
-CUSTOM_TYPE = object()
-"""
-When given as a key, indicates that the value denotes a custom type for the object.
-"""
-
-NOTHING_T = Literal[NOTHING]
-
-IGNORE_T = Literal[IGNORE]
-
-CUSTOM_TYPE_T = Literal[CUSTOM_TYPE]
-
-KEY_T = Union[str, NOTHING_T]
-
-POSSIBLE_JSON_T = Union[JSON_T, IGNORE_T]
-
-
-def escape(value: JSON_T = None) -> JSON_T:
-    """If the value is a string beginning with '$', replace it with '$$'.
-
-    Recursively descends into child objects.
-    """
-    if isinstance(value, str) and value.startswith("$"):
-        return "$" + value
-    if isinstance(value, MutableMapping):
-        return {k: escape(v) for k, v in value.items()}
-    return value
-
-
-def format_unset(default: JSON_T = NOTHING, max_default_len=35) -> str:
-    """Return '$#', optionally with a parenthesized default value added after a space."""
-    if default is not NOTHING:
-        default = brief(repr(default), max_len=max_default_len)
-        return f"$# ({default})"
-    return f"$#"
-
-
-class RemoteReferenceMixin(ABC):
-    def __init__(self, *, target: Path, **kwargs) -> None:
-        super().__init__(**kwargs)
-        self.target = target
-
-    @property
-    def reference_str(self) -> str:
-        """The absolute or relative control sequence string used to reference the object."""
-        if self.target.is_absolute():
-            return f"{MARKER}/{self.target}"
-        return f"{MARKER}./{self.target}"
-
-    @abstractmethod
-    def expand(self) -> JSON_T:
-        """Return the referenced data."""
-
-    def write(
-        self,
-        serializer: "type[JSONEncoder]" = JSONEncoder,
-        location: Optional[Path] = None,
-        write_trailing_newline: bool = True,
-        **kwargs,
-    ) -> None:
-        """Serialize the contents of this mapping to its target."""
-        target = self.target
-        if not target.is_absolute():
-            assert (
-                location
-            ), f"Cannot write to relative path '{target}' without a location specified."
-            target = location / target
-        target.parent.mkdir(exist_ok=True, parents=True)
-        with open(target, "w") as f:
-            json.dump(self.expand(), f, cls=serializer, location=location, **kwargs)
-            if write_trailing_newline:
-                f.write("\n")
-
-
-class RemoteMapping(RemoteReferenceMixin, WrappedDict):
-    """A dict-like reference to a JSON mapping (object) stored in another file."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-    def expand(self) -> dict:
-        return self._dict
-
-
-def expand_mappings(mapping: MutableMapping) -> dict:
-    """Recursively expand any remote mappings within `mapping`."""
-    ret = dict()
-    for keys, val in walk_dict(mapping):
-        if isinstance(val, RemoteMapping):
-            set_nested(ret, keys, val.expand())
-        else:
-            set_nested(ret, keys, val)
-    return ret
-
-
-class VJSONDecoder(JSONDecoder):
-    """A custom JSON decoder which supports references to objects in other files.
-
-    Control sequences are two-character sequences which must occur at the beginning of an encoded string,
-        and are defined as follows:
-
-    `$$`: A single `$` character. Valid as either a key or value.
-    `$.`: Denotes the beginning of a relative path to anothor .json or .v.json spec whose contents are to be substituted
-          in its place. The spec must have a single object at its root. Valid only as a value.
-    `$/`: Similar to above, but with an absolute path. Valid only as a value.
-    `$#`: When given as a value, optionally followed by any sequence of characters (which is ignored), signifies that
-          a keypair should be entirely ignored by the decoder.
-    `$t`: Specifies the unique type-identifier of a registered VJSON type to which the containing object should
-          be deserialized.
-
-    The class can be instantiated with or without support for the relative path control sequence.
-
-    Keyword arguments:
-        location: The path to the file being decoded (to allow expansion of relative path references).
-                  Defaults to `None` (attempting to load a `$.` sequence will raise `ValueError`
-                  in this case).
-
-    All other positional and keyword arguments are passed to `JSONDecoder.__init__()`.
-    The `object_hook` keyword cannot be set since it is used internally.
-    """
-
-    def __init__(self, location: Optional[Path] = None, **kwargs) -> None:
-        super().__init__(object_hook=self.object_hook, **kwargs)
-        self.location = location
-        self._params = kwargs.copy()
-        self._csm = {
-            MARKER * 2: self.expand_escaped,
-            f"{MARKER}.": self.expand_relative,
-            f"{MARKER}/": self.expand_absolute,
-            f"{MARKER}#": self.expand_default,
-            f"{MARKER}t": self.expand_custom_type,
-        }  # Stands for "control sequence mapping".
-
-    def params(self) -> dict:
-        """Return a dictionary of all keyword arguments used to instantiate this object apart from `location`."""
-        return self._params.copy()
-
-    def object_hook(self, obj: dict) -> dict:
-        gen_expanded = (self.expand(pair) for pair in obj.items())
-        ret = {k: v for k, v in gen_expanded if v is not IGNORE}
-        custom_type = ret.pop(CUSTOM_TYPE, None)
-        if custom_type:
-            return custom_type.from_object(ret)
-        return ret
-
-    def expand_custom_type(
-        self, value: JSON_T, key: KEY_T = NOTHING
-    ) -> "tuple[CUSTOM_TYPE_T, VJSONSerializableMixin]":
-        assert key == f"{MARKER}t"
-        return CUSTOM_TYPE, VJSONSerializableMixin.by_type_id(value)
-
-    def expand_default(self, value: JSON_T, key: KEY_T = NOTHING) -> IGNORE_T:
-        """Return `IGNORE`."""
-        assert key is NOTHING, "'$#' cannot be expanded in a key."
-        assert value.startswith(f"{MARKER}#")
-        return IGNORE
-
-    def expand_escaped(
-        self, value: JSON_T, key: KEY_T = NOTHING
-    ) -> "Union[VJSON_T, tuple[KEY_T, VJSON_T]]":
-        """Expand an escaped string. Works for either values or keys with values.
-
-        If both are given, expands the key and returns a tuple of the expanded `key` and unchanged `value`.
-        """
-        if key is NOTHING:
-            return value[1:]
-        return key[1:], value
-
-    def load_custom_or_remote(self, path: Path) -> "VJSON_T":
-        """Load the file at `path` as either a custom object or a `RemoteMapping`.
-
-        If the file doesn't exist, it is created and initialized with an empty
-        object.
-
-        The file must contain a single JSON object.
-        """
-        if not path.exists():
-            path.parent.mkdir(exist_ok=True, parents=True)
-            with open(path, "w") as f:
-                json.dump(dict(), fp=f)
-            return RemoteMapping(target=path)
-        with open(path, "r") as f:
-            data = json.load(f, cls=type(self), location=path.parent, **self.params())
-        if isinstance(data, VJSONSerializableMixin):
-            return data
-        return RemoteMapping(target=path, **data)
-
-    def expand_absolute(self, value: str, key: KEY_T = NOTHING) -> RemoteMapping:
-        """Load the absolute path given in `value`. The path cannot be in the key.
-
-        If the file doesn't exist, it is created and initialized with an empty object.
-
-        It is an error to set `key`. The file pointed to at `value` is must contain
-        a single JSON object.
-        """
-        assert key is NOTHING, "Absolute path references are not supported in keys."
-        path = Path(value[2:])
-        return self.load_custom_or_remote(path)
-
-    def expand_relative(self, value: str, key: KEY_T = NOTHING) -> RemoteMapping:
-        f"""Load the given relative `path`."""
-        assert key is NOTHING, "Relative path references are not supported in keys."
-        if not self.location:
-            raise ValueError("This decoder does not support relative path references.")
-        return self.expand_relative_to(self.location, value, key=key)
-
-    def expand_relative_to(
-        self, location: Path, value: JSON_T, key: KEY_T = NOTHING
-    ) -> RemoteMapping:
-        """Load the path given in `value` relative to `location`. The path cannot be in the key.
-
-        If the file doesn't exist, it is created and initialized with an empty object.
-
-        It is an error to set `key`. The file pointed to by the value must contain a JSON object.
-        """
-        assert key is NOTHING, "Relative path references are not supported in keys."
-        relpath = value[3:]
-        path = location / relpath
-        return self.load_custom_or_remote(path)
-
-    def expand(self, pair: "tuple[str, JSON_T]") -> "tuple[str, JSON_T]":
-        """Expand the (key, value) pair as appropriate.
-
-        First checks the key for a control sequence, then the value.
-
-        Control sequence expansion methods should return only a value if only the `value`
-        parameter is provided. Likewise, they should return a tuple of both the expanded
-        key and the original value if provided with both the `key` and `value` parameters.
-        """
-        k, v = pair
-        k2 = k[:2]
-        if k2 in self._csm:
-            k, v = self._csm[k2](v, key=k)
-        if isinstance(v, str):
-            v2 = v[:2]
-            if v2 in self._csm:
-                v = self._csm[v2](v)
-        elif isinstance(v, dict):
-            v = self.object_hook(v)
-        return k, v
-
-
-class VJSONEncoder(JSONEncoder):
-    """Serializer counterpart to `VJSONDecoder`.
-
-    See `VJSONDecoder` for further information.
-    """
-
-    def __init__(
-        self,
-        location: Optional[Path] = None,
-        **kwargs,
-    ) -> None:
-        """
-        **kwargs:
-            location (Path): The base path to use when expanding relative references.
-            write_remote_mappings (bool): Whether to update the files corresponding to
-                remote references during serialization.
-            expand_remote_mappings (bool): If True, write the content of the files corresponding
-                to remote references directly in the output file.
-            - Remaining arguments are passed to JSONEncoder.__init__().
-        """
-        self._params = kwargs.copy()
-        self.write_remote_mappings = kwargs.pop("write_remote_mappings")
-        self.expand_remote_mappings = kwargs.pop("expand_remote_mappings")
-        self.location = location
-        super().__init__(**kwargs)
-
-    def params(self) -> dict:
-        """Return a dictionary of all keyword arguments used to instantiate this object apart from `location`."""
-        return self._params.copy()
-
-    def default(self, o: "VJSON_T") -> JSON_T:
-        if isinstance(o, RemoteReferenceMixin):
-            if self.write_remote_mappings:
-                o.write(type(self), self.location, **self.params())
-            if self.expand_remote_mappings:
-                return o.expand()
-            return o.reference_str
-        if isinstance(o, VJSONSerializableMixin):
-            return o.to_object()
-        return super().default(o)
-
-
-def load(source: Path, location: Path = None, **kwargs) -> MutableMapping:
-    """Load the object stored at `source` using a VJSONDecoder."""
-    if location is None:
-        location = source.parent
-    with open(source, "r") as f:
-        return json.load(f, cls=VJSONDecoder, location=location, **kwargs)
-
-
-def loads(s: str, **kwargs) -> MutableMapping:
-    """Load the given VJSON-formatted string into a dict."""
-    return json.loads(s, cls=VJSONDecoder, **kwargs)
-
-
-def dump(
-    obj: "MutableMapping[str, VJSON_T]",
-    dest: Path,
-    location: Path = None,
-    write_remote_mappings: bool = True,
-    expand_remote_mappings: bool = False,
-    write_trailing_newline: bool = True,
-    indent: Optional[int] = 4,
-    **kwargs,
-) -> None:
-    """Dump the given object into the file specified by `dest` using a VJSONEncoder."""
-    if location is None:
-        location = dest.parent
-    with open(dest, "w") as f:
-        json.dump(
-            obj,
-            f,
-            cls=VJSONEncoder,
-            indent=indent,
-            location=location,
-            write_remote_mappings=write_remote_mappings,
-            expand_remote_mappings=expand_remote_mappings,
-            **kwargs,
-        )
-        if write_trailing_newline:
-            f.write("\n")
-
-
-def dumps(
-    obj: "MutableMapping[str, VJSON_T]",
-    location: Path = None,
-    write_remote_mappings: bool = True,
-    expand_remote_mappings: bool = False,
-    indent: Optional[int] = 4,
-    **kwargs,
-) -> str:
-    """Dump the given object as a VJSON-formatted string."""
-    return json.dumps(
-        obj,
-        cls=VJSONEncoder,
-        indent=indent,
-        location=location,
-        write_remote_mappings=write_remote_mappings,
-        expand_remote_mappings=expand_remote_mappings,
-        **kwargs,
-    )
-
-
-class VJSONSerializableMixin:
-    """Mixin for types which define `to_object` and `from_object` methods.
-
-    When subclassing, the `type_id` class attribute may be provided as
-    a name to use in the serial representation's `"type"` attribute.
-
-    This associates the string with the new type, thus allowing objects to be reconstructed
-    "automagically" from their serial format.
-
-    Inheriting from this mixin is currently the only way to create a custom type which is
-    automatically (de)serializable.
-    """
-
-    named_types = dict()
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._target = None
-
-    def __init_subclass__(cls, **kwargs) -> None:
-        type_id = getattr(cls, "type_id", None)
-        if type_id is not None:
-            VJSONSerializableMixin.named_types[type_id] = cls
-        return super().__init_subclass__(**kwargs)
-
-    @abstractmethod
-    def to_object(self) -> "dict[str, VJSON_T]":
-        """Return a dictionary representation of this object.
-
-        The attributes of this dictionary must constitute the complete set of constructor
-        keyword parameters needed to recreate this instance (this mixin could be refactored
-        to i.e. inherit from `dataclass.dataclass` in the future).
-
-        Implementations must start by calling `super().to_object()` and then
-        update the resulting dictionary with their serializable data. This way,
-        special attributes added by `VJSONSerializableMixin` is preserved, and the
-        type of mapping to use (`RemoteMapping` or `dict`) can be determined dynamically.
-        """
-        if self._target is not None:
-            mapping = RemoteMapping(target=self._target)
-        else:
-            mapping = dict()
-        mapping.update({f"{MARKER}t": self.type_id})
-        return mapping
-
-    @classmethod
-    def from_object(cls, obj: "dict[str, VJSON_T]") -> "VJSONSerializableMixin":
-        """Deserialize the given object and return the new instance of this type."""
-        instance = cls(**obj)
-        if isinstance(obj, RemoteMapping):
-            instance._target = obj.target
-        return instance
-
-    @classmethod
-    def by_type_id(cls, type_id: str) -> "VJSONSerializableMixin":
-        """Return the class associated with the given type id."""
-        return cls.named_types[type_id]
-
-    def save(self, to: Path, **kwargs) -> None:
-        """Save this object to a VJSON file at the given path.
-
-        This implicitly saves any child objects which are provided by `self.to_object`
-        so it should only be called once when saving hierarchical types
-        (on the object at the top of the hierarchy).
-
-        Extra keyword arguments are passed to `vjson.dump()`.
-        """
-        obj = self.to_object()
-        location = to.parent
-        location.mkdir(exist_ok=True, parents=True)
-        dump(obj, dest=str(to.resolve()), location=location, **kwargs)
-
-    @classmethod
-    def load(cls, from_: Path) -> "VJSONSerializableMixin":
-        """Load the object in the vjson file at the given path."""
-        return load(source=from_, location=from_.parent)
-
-
-VJSON_T = Union[JSON_T, RemoteMapping, VJSONSerializableMixin]
-"""A type which can be represented as a string in the .v.json format."""
